Fix integer division in ReviewStats.getPercentageForRating (CALCULATION-1)
Changed return type from Int to Double to preserve decimal precision.
Example: 1 out of 3 reviews now returns 33.333... instead of 33.
Also updated RatingDistributionRow caller to accept Double.

+++ feature/reviews/src/commonMain/kotlin/com/aggregateservice/feature/reviews/domain/model/ReviewStats.kt
@@ -25,10 +25,10 @@ data class ReviewStats(
     /**
      * Percentage of reviews for a given rating.
      */
-    fun getPercentageForRating(rating: Int): Int {
-        if (totalReviews == 0) return 0
+    fun getPercentageForRating(rating: Int): Double {
+        if (totalReviews == 0) return 0.0
         val count = ratingDistribution[rating] ?: 0
-        return (count * 100) / totalReviews
+        return (count * 100.0) / totalReviews
     }
 
     /**
